third_party/blink/web_tests/fast/xmlhttprequest/xmlhttprequest-get-expected.txt

GET test
responseText
<?xml version="1.0"?>
<!DOCTYPE doc [
<!ATTLIST d id ID #IMPLIED>
]>
<doc>
  <foo xmlns="foobar">One</foo> <x:bar xmlns:x="barfoo">Two</x:bar>
  <d id="id3">Three</d>
</doc>

responseXML serialized
<?xml version="1.0"?><!DOCTYPE doc><doc>
  <foo xmlns="foobar">One</foo> <x:bar xmlns:x="barfoo">Two</x:bar>
  <d id="id3">Three</d>
</doc>
getAllResponseHeaders()
content-type: text/xml
last-modified: (date)

status
200
statusText
OK
readyState
4
Event information
Event object: [object ProgressEvent]
Event properties:
AT_TARGET : '2'
BUBBLING_PHASE : '3'
CAPTURING_PHASE : '1'
NONE : '0'
bubbles : 'false'
cancelBubble : 'false'
cancelable : 'false'
composed : 'false'
composedPath : 'function composedPath() { [native code] }'
currentTarget : '[object XMLHttpRequest]'
defaultPrevented : 'false'
eventPhase : '2'
initEvent : 'function initEvent() { [native code] }'
isTrusted : 'true'
lengthComputable : 'true'
loaded : '0'
preventDefault : 'function preventDefault() { [native code] }'
returnValue : 'true'
srcElement : '[object XMLHttpRequest]'
stopImmediatePropagation : 'function stopImmediatePropagation() { [native code] }'
stopPropagation : 'function stopPropagation() { [native code] }'
target : '[object XMLHttpRequest]'
total : '0'
type : 'load'
